pyproject.toml
[tool.poetry]
name = "centraljersey"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.5"
ipykernel = "^6.20.2"
PyYAML = "^6.0"
pandas = "^1.5.3"
geopandas = "^0.12.2"
folium = "^0.14.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.20.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
